move tests into own module
